# docs/openapi/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "HealthResponse": {
        "additionalProperties": false,
        "properties": {
          "degraded_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Degraded Reasons",
            "type": "array"
          },
          "dependencies": {
            "additionalProperties": true,
            "title": "Dependencies",
            "type": "object"
          },
          "real_service_acceptance": {
            "title": "Real Service Acceptance",
            "type": "boolean"
          },
          "runtime": {
            "title": "Runtime",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "runtime",
          "real_service_acceptance"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Enterprise RAG Knowledge Base",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/capabilities": {
      "get": {
        "operationId": "capabilities_api_v1_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Capabilities Api V1 Capabilities Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Capabilities",
        "tags": [
          "health"
        ]
      }
    },
...
    "/status/acceptance": {
      "get": {
        "operationId": "acceptance_status_status_acceptance_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Acceptance Status",
        "tags": [
          "health"
        ]
      }
    },
    "/status/degraded": {
      "get": {
        "operationId": "degraded_status_status_degraded_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Degraded Status",
        "tags": [
          "health"
        ]
      }
    }
  }
}